[driver/labjack] restructure code
