+++ driver/labjack/factory.cpp
@@ -11,7 +11,8 @@
 #include "glog/logging.h"
 #include "driver/labjack/labjack.h"
 #include "driver/labjack/scanner.h"
-#include "driver/labjack/task.h"
+#include "driver/labjack/reader.h"
+#include "driver/labjack/writer.h"
 
 std::pair<std::unique_ptr<task::Task>, bool> labjack::Factory::configure_task(
         const std::shared_ptr<task::Context> &ctx,
