+++ driver/labjack/reader_test.cpp
@@ -10,7 +10,6 @@
 #include <stdio.h>
 
 #include "client/cpp/synnax.h"
-#include "driver/labjack/task.h"
 #include "driver/labjack/reader.h"
 #include "driver/labjack/writer.h"
 #include "driver/labjack/scanner.h"
