+++ driver/labjack/test_labjack.cpp
@@ -288,9 +288,6 @@ int multi_ain(){
     return LJME_NOERROR;
 }
 
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////////////////////////STREAMING
-
 void HardcodedConfigureStream(int handle)
 {
     const int STREAM_TRIGGER_INDEX = 0;
@@ -546,11 +543,6 @@ int PrintErrors(){
 
     return LJME_NOERROR;
 }
-
-
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////// END OF STREAMING
 
 int main() {
 //    return PrintErrors();
