+++ driver/labjack/writer_sink.cpp
